refactor: combine delivery modules
