+++ embrace-android-core/src/main/kotlin/io/embrace/android/embracesdk/internal/injection/DeliveryModule.kt
@@ -1,7 +1,25 @@
 package io.embrace.android.embracesdk.internal.injection
 
 import io.embrace.android.embracesdk.internal.comms.delivery.DeliveryService
+import io.embrace.android.embracesdk.internal.delivery.caching.PayloadCachingService
+import io.embrace.android.embracesdk.internal.delivery.execution.RequestExecutionService
+import io.embrace.android.embracesdk.internal.delivery.intake.IntakeService
+import io.embrace.android.embracesdk.internal.delivery.resurrection.PayloadResurrectionService
+import io.embrace.android.embracesdk.internal.delivery.scheduling.SchedulingService
+import io.embrace.android.embracesdk.internal.delivery.storage.PayloadStorageService
 
+/**
+ * Contains dependencies that are required for delivering code to Embrace servers.
+ *
+ * Types are nullable because folks who are only exporting via OTel exporters do not need this
+ * functionality.
+ */
 interface DeliveryModule {
     val deliveryService: DeliveryService
+    val intakeService: IntakeService?
+    val payloadResurrectionService: PayloadResurrectionService?
+    val payloadCachingService: PayloadCachingService?
+    val payloadStorageService: PayloadStorageService?
+    val requestExecutionService: RequestExecutionService?
+    val schedulingService: SchedulingService?
 }
