+++ embrace-android-core/src/main/kotlin/io/embrace/android/embracesdk/internal/injection/DeliveryModuleSupplier.kt
@@ -1,22 +1,29 @@
 package io.embrace.android.embracesdk.internal.injection
 
-import io.embrace.android.embracesdk.internal.comms.api.ApiService
-
 /**
  * Function that returns an instance of [DeliveryModule]. Matches the signature of the constructor for [DeliveryModuleImpl]
  */
 typealias DeliveryModuleSupplier = (
+    configModule: ConfigModule,
     initModule: InitModule,
+    workerThreadModule: WorkerThreadModule,
+    coreModule: CoreModule,
     storageModule: StorageModule,
-    apiService: ApiService?,
+    essentialServiceModule: EssentialServiceModule,
 ) -> DeliveryModule
 
 fun createDeliveryModule(
+    configModule: ConfigModule,
     initModule: InitModule,
+    workerThreadModule: WorkerThreadModule,
+    coreModule: CoreModule,
     storageModule: StorageModule,
-    apiService: ApiService?,
+    essentialServiceModule: EssentialServiceModule,
 ): DeliveryModule = DeliveryModuleImpl(
+    configModule,
     initModule,
+    workerThreadModule,
+    coreModule,
     storageModule,
-    apiService,
+    essentialServiceModule
 )
