+++ embrace-android-sdk/src/integrationTest/kotlin/io/embrace/android/embracesdk/IntegrationTestRule.kt
@@ -99,7 +99,7 @@ internal class IntegrationTestRule(
                 coreModuleSupplier = { _, _ -> overriddenCoreModule },
                 workerThreadModuleSupplier = { _ -> overriddenWorkerThreadModule },
                 androidServicesModuleSupplier = { _, _, _ -> overriddenAndroidServicesModule },
-                deliveryModuleSupplier = { _, _, _ -> overriddenDeliveryModule },
+                deliveryModuleSupplier = { _, _, _, _, _, _ -> overriddenDeliveryModule },
                 anrModuleSupplier = { _, _, _, _ -> fakeAnrModule },
                 nativeFeatureModuleSupplier = { _, _, _, _, _, _, _, _, _ -> fakeNativeFeatureModule }
             )
